[IMP] Add queue_job.keep_context ir.config_parameter
When this parameter is set, queue_job always preserves the entire
context. This honors the principle of least surprise, in that a
developer can easily convert a record.method() call to
record.with_delay().method() with the expectation that it will
actually execute the same, simply at a later time.